requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
accessible-pygments=0.0.5=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.11.7=pypi_0
aiosignal=1.3.1=pypi_0
alabaster=1.0.0=pypi_0
albucore=0.0.20=pypi_0
albumentations=1.4.21=pypi_0
annotated-types=0.7.0=pypi_0
anomalib=1.2.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
attrs=24.2.0=pypi_0
babel=2.16.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.2.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.9.24=haa95532_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
defusedxml=0.7.1=pypi_0
docstring-parser=0.16=pypi_0
docutils=0.21.2=pypi_0
einops=0.8.0=pypi_0
eval-type-backport=0.2.0=pypi_0
fastjsonschema=2.20.0=pypi_0
filelock=3.16.1=pypi_0
fonttools=4.55.0=pypi_0
freia=0.2=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.10.0=pypi_0
ftfy=6.3.1=pypi_0
huggingface-hub=0.26.2=pypi_0
idna=3.10=pypi_0
imageio=2.36.0=pypi_0
imagesize=1.4.1=pypi_0
imgaug=0.4.0=pypi_0
importlib-resources=6.4.5=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jsonargparse=4.34.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
kiwisolver=1.4.7=pypi_0
kornia=0.7.4=pypi_0
kornia-rs=0.1.7=pypi_0
lazy-loader=0.4=pypi_0
libffi=3.4.4=hd77b12b_1
lightning=2.4.0=pypi_0
lightning-utilities=0.11.9=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.2=pypi_0
mdit-py-plugins=0.4.2=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
myst-parser=4.0.0=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
nbsphinx=0.9.5=pypi_0
networkx=3.4.2=pypi_0
numpy=1.24.3=pypi_0
omegaconf=2.3.0=pypi_0
open-clip-torch=2.29.0=pypi_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openssl=3.0.15=h827c3e9_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pandoc=2.4=pypi_0
pandocfilters=1.5.1=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py311haa95532_0
platformdirs=4.3.6=pypi_0
plumbum=1.9.0=pypi_0
ply=3.11=pypi_0
propcache=0.2.0=pypi_0
pydantic=2.10.1=pypi_0
pydantic-core=2.27.1=pypi_0
pydata-sphinx-theme=0.16.0=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.2.0=pypi_0
python=3.11.10=h4607a30_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-graphviz=0.20.3=pypi_0
pytorch-lightning=2.4.0=pypi_0
pytz=2024.2=pypi_0
pywin32=308=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=pypi_0
referencing=0.35.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=13.9.4=pypi_0
rich-argparse=1.6.0=pypi_0
rpds-py=0.21.0=pypi_0
safetensors=0.4.5=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
setuptools=75.1.0=py311haa95532_0
shapely=2.0.6=pypi_0
simsimd=6.1.1=pypi_0
six=1.16.0=pypi_0
snowballstemmer=2.2.0=pypi_0
soupsieve=2.6=pypi_0
sphinx=8.1.3=pypi_0
sphinx-autodoc-typehints=2.5.0=pypi_0
sphinx-book-theme=1.1.3=pypi_0
sphinx-copybutton=0.5.2=pypi_0
sphinx-design=0.6.1=pypi_0
sphinxcontrib-applehelp=2.0.0=pypi_0
sphinxcontrib-devhelp=2.0.0=pypi_0
sphinxcontrib-htmlhelp=2.1.0=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=2.0.0=pypi_0
sphinxcontrib-serializinghtml=2.0.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
stringzilla=3.10.10=pypi_0
sympy=1.13.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.9.20=pypi_0
timm=1.0.11=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.14=h0416ee5_0
torch=2.5.1=pypi_0
torchmetrics=1.6.0=pypi_0
torchvision=0.20.1=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.67.0=pypi_0
traitlets=5.14.3=pypi_0
typeshed-client=2.7.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
wcwidth=0.2.13=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.44.0=py311haa95532_0
xz=5.4.6=h8cc25b3_1
yarl=1.18.0=pypi_0
zlib=1.2.13=h8cc25b3_1
